requirements.txt
mediapipe
PyOpenGL